added profile for easier save management

+++ functionForSave.py
@@ -9,34 +9,85 @@ for i in os.listdir(folderLocation) :
         CurrentSavePath = folderLocation+"/"+i
         break
 
-def importSave(Name):
-    NewSavePath = AllSavePath +"/"+Name
+def ImportSave(Name,Profile):
+    if Profile == "":
+        return False
+    NewSavePath = AllSavePath +"/"+Profile+"/"+Name
     if os.path.exists(NewSavePath) :        
         os.chmod(NewSavePath,stat.S_IWUSR)
         shutil.rmtree(NewSavePath)
     shutil.copytree(CurrentSavePath,NewSavePath,dirs_exist_ok=True)
+    return True
 
-def RemoveSave(Name):
-    RemovedSavePath = AllSavePath +"/"+Name
+def DuplicateSave(Name,NameOfCopy,Profile):
+    if Profile == "":
+        return False
+    OldSavePath = AllSavePath +"/"+Profile+"/"+Name
+    NewSavePath = AllSavePath +"/"+Profile+"/"+NameOfCopy
+    if os.path.exists(NewSavePath) :        
+        os.chmod(NewSavePath,stat.S_IWUSR)
+        shutil.rmtree(NewSavePath)
+    shutil.copytree(OldSavePath,NewSavePath,dirs_exist_ok=True)
+    return True
+
+def RemoveSave(Name,Profile):
+    if Profile == "":
+        return False
+    RemovedSavePath = AllSavePath +"/"+Profile+"/"+Name
     os.chmod(RemovedSavePath,stat.S_IWUSR)
     shutil.rmtree(RemovedSavePath)
+    return True
 
-def RenameSave(OldName,NewName):
-    OldNamePath = AllSavePath +"/"+OldName
-    NewNamePath = AllSavePath +"/"+NewName
+def RenameSave(OldName,NewName,Profile):
+    if Profile == "":
+        return False
+    OldNamePath = AllSavePath +"/"+Profile+"/"+OldName
+    NewNamePath = AllSavePath +"/"+Profile+"/"+NewName
     os.rename(OldNamePath,NewNamePath)
+    return True
 
-def LoadSave(Name):
-    LoadedSavePath = AllSavePath +"/"+Name
+def LoadSave(Name,Profile):
+    if Profile == "":
+        return False
+    LoadedSavePath = AllSavePath +"/"+Profile+"/"+Name
     if os.path.exists(CurrentSavePath):
         for i in os.listdir(CurrentSavePath):
             os.remove(CurrentSavePath +"/"+i)
     shutil.copytree(LoadedSavePath,CurrentSavePath,dirs_exist_ok=True)
 
-def GetListOfSave():
+def GetListOfSave(Profile):
+    return os.listdir(AllSavePath+"/"+Profile)
+
+def GetListOfProfile():
     return os.listdir(AllSavePath)
 
+def CreateProfile(Profile):
+    if Profile == "":
+        return False
+    NewProfilePath = AllSavePath +"/"+Profile
+    if os.path.exists(NewProfilePath) :
+        return "Profile already exist"
+    os.makedirs(NewProfilePath,exist_ok=True)
 
+def DeleteProfile(Profile):
+    if Profile == "":
+        return False
+    RemovedprofilePath = AllSavePath +"/"+Profile
+    for root,dirs,files in os.walk(RemovedprofilePath):
+        os.chmod(root,stat.S_IWUSR)
+        for d in dirs:
+            dirsPath = RemovedprofilePath + "/" + d
+            os.chmod(dirsPath,stat.S_IWUSR)
+    shutil.rmtree(RemovedprofilePath)
 
+def DuplicateProfile(Profile, NameOfCopy):
+    profilePath = AllSavePath +"/"+Profile
+    CopyPath = AllSavePath +"/"+NameOfCopy
+    shutil.copytree(profilePath,CopyPath,dirs_exist_ok=True)
+
+def RenameProfile(Profile, NewName):
+    OldProfilePath = AllSavePath +"/"+Profile
+    NewProfilePath = AllSavePath +"/"+NewName
+    os.rename(OldProfilePath,NewProfilePath)
